Fix support for non-default kubelet root directory
It turns out that the container-local path also needs to be updated to
match what the path is on the host. Intuitively it seemed that updating
the container-local path would have broken assumptions in the provider,
but *counter-intuitively* it turns out that the mismatch between
container and host path strings breaks the CSI driver stack.

The breakage presents as the upstream CSI driver (not the AWS provider)
being unable to mount the secrets volumes in pods.