Continue to fetch secret value even if describe access denied
Currently, if `DescribeSecret` is not allowed for one of the secrets,
the entire `Mount` request and the auto rotation would fail.
But sometimes secrets might be created with only `GetSecret` allowed,
which is still enough permission to generate a new secret.
It would be great to still try fetching secret value even if describe
failed because of AccessDenied.