Add documentation about Helm chart CSI driver installation. (#495)
*Description of changes:*

This PR adds documentation clarifying that the CSI driver installation
step can be skipped when using Helm, as the chart automatically installs
a compatible version of the Secrets Store CSI driver as a dependency. It
also notes that this automatic installation can be disabled by setting
`secrets-store-csi-driver.install=false`

PR:
https://github.com/aws/secrets-store-csi-driver-provider-aws/pull/384

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.