feat: add EKS addon field to user agent middleware (#502)
*Issue #, if available:*
N/A

*Description of changes:*
A new field in `values.yaml`, `eks-addon-version`, will be set to the
addon version in the EKS addon copy of the file. This will cause the EKS
addon version to be appended to the User Agent string.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Signed-off-by: Simon Marty <simon.marty0@gmail.com>
Co-authored-by: Simon Marty <simon.marty0@gmail.com>